+++ lib/modules/clients/domain/domain.dart
@@ -1,6 +1,7 @@
 export 'model/address.dart';
 export 'model/client.dart';
 export 'model/contact.dart';
+export 'model/states.dart';
 export 'repository/client_repository.dart';
 export 'usecase/usecase.dart';
 export 'dto/listing_client_dto.dart';
